+++ src/bytecode.rs
@@ -113,6 +113,7 @@ pub enum Op {
         bindings: usize,
     },
     MatchFailure,
+    Throw,
     Recur(usize),
     Return,
 }